feat(clash): 重构订阅信息处理逻辑
- 新增 SubscriptionUserinfo 类封装订阅信息
- 在 ClashData 中添加 subscriptionUserinfo 字段
- 修改 EditProfileActivity 和 NewProfileActivity 以支持新的订阅信息处理方式
- 更新 SubscriptionUserinfoManager 以使用 Long 类型的 profileId

+++ app/src/main/java/io/nekohasekai/sfa/clash/SubscriptionUserinfo.kt
@@ -6,14 +6,14 @@ import kotlinx.parcelize.Parcelize
 
 object SubscriptionUserinfoManager {
     private val mmkv: MMKV = MMKV.mmkvWithID("subscription_userinfo", MMKV.MULTI_PROCESS_MODE)
-    fun setUserinfo(profileId: Int, userinfo: SubscriptionUserinfo) {
+    fun setUserinfo(profileId: Long, userinfo: SubscriptionUserinfo) {
         mmkv.encode("profile_$profileId", userinfo)
     }
-    fun getUserinfo(profileId: Int): SubscriptionUserinfo? {
+    fun getUserinfo(profileId: Long): SubscriptionUserinfo? {
         return mmkv.decodeParcelable("profile_$profileId", SubscriptionUserinfo::class.java)
     }
 
-    fun remove(profileId: Int) {
+    fun remove(profileId: Long) {
         mmkv.removeValueForKey("profile_$profileId")
     }
     fun clear() {
